data/socketserver/requirements.txt
pyOpenSSL
python-dotenv
